fix: 各種ライブラリのアップデート及びREADMEの更新、418ハイドレーションエラー対策とNextの将来的な仕様変更に備えた調整.
